requirements.txt
grpcio==1.58.0
joblib==1.3.2
numpy==1.25.2
protobuf==4.24.3
scikit-learn==1.3.0
scipy==1.11.2
threadpoolctl==3.2.0
